Python updated to version 3.13
* updated the python version to latest

* Changed the python version in all workflows

* Added azure identity package to requirements

* Changed the authentication type

* Changed python version to 3.13 and enabled manual trigger

* Changed vnet name

* Changed vnet create function

* Changed subnets create function

* Changed subnet name to subnet list

* Fixed subnet name

* Changed the deletion of vnet resource group

* Added existing VPC condition

+++ healthcheck/healthcheck_webapps_refarch.py
@@ -35,12 +35,13 @@ def main(tenant_id_arg, client_id_arg, client_secret_arg, subscription_id_arg, u
         resource_name_vnet = 'vnet_resource_group'
         # Deploy a resource group with a virtual network and specified number of subnets
         try:
-            subnet_name, vnet_name = DeployOp.create_vnet(credentials,
+            subnet_name_list, vnet_name = DeployOp.create_vnet(credentials,
                                                             subscription_id,
                                                             location,
                                                             subnets_cidr,
                                                             resource_name_vnet,
                                                             vnet_cidr)
+            subnet_name = subnet_name_list[0]
             network_client = NetworkManagementClient(credentials, subscription_id)
             # Add service endpoint to subnets
             subnet = network_client.subnets.get(resource_name_vnet, vnet_name, subnet_name)
@@ -49,7 +50,7 @@ def main(tenant_id_arg, client_id_arg, client_secret_arg, subscription_id_arg, u
             subnet.service_endpoints.append(
                 ServiceEndpointPropertiesFormat(service='Microsoft.Storage')
             )
-            updated_subnet = network_client.subnets.create_or_update(
+            updated_subnet = network_client.subnets.begin_create_or_update(
                 resource_name_vnet,
                 vnet_name,
                 subnet_name,
@@ -73,7 +74,7 @@ def main(tenant_id_arg, client_id_arg, client_secret_arg, subscription_id_arg, u
             "Deploy to New or Existing Virtual Network": "existing",
             "Name of Virtual Network Where MATLAB Web App Server Will Be Deployed": vnet_name,
             "Virtual Network CIDR Range": vnet_cidr,
-            "Name of Subnet for MATLAB Web App Server": subnet_name[0],
+            "Name of Subnet for MATLAB Web App Server": subnet_name,
             "Server Subnet CIDR Range": subnets_cidr[0],
             "Specify Private IP Address to VM Hosting MATLAB Web App Server": '10.1.0.4',
             "Resource Group Name Of Virtual Network": resource_name_vnet,
@@ -84,7 +85,10 @@ def main(tenant_id_arg, client_id_arg, client_secret_arg, subscription_id_arg, u
     res = requests.get(
         f"https://github.com/mathworks-ref-arch/{ref_arch_name}/tree/main/releases/"
     )
-    latest_releases = [re.findall("releases/(R\d{4}[ab]\\b)", res.text)[-1], re.findall("releases/(R\d{4}[ab]\\b)", res.text)[-2]]
+    latest_releases = [
+        re.findall(r"releases/(R\d{4}[ab]\b)", res.text)[-1],
+        re.findall(r"releases/(R\d{4}[ab]\b)", res.text)[-2]
+    ]
     for i in range(2):
         matlab_release = latest_releases[i]
         print("Testing Health Check Release: " + matlab_release + "\n")
@@ -114,15 +118,17 @@ def main(tenant_id_arg, client_id_arg, client_secret_arg, subscription_id_arg, u
                 print("Deleted the deployment which is deployed using existing virtual network:-",ct)
                 # Wait for above deployment deletion
                 time.sleep(900)
-                # Delete deployment with virtual network
-                DeployOp.delete_resourcegroup(credentials, subscription_id, resource_name_vnet)
-                ct = datetime.datetime.now()
-                print("Deleted the deployment which contains the virtual network:-",ct)
              else:
                 # Delete the deployment
                 deployment_deletion = DeployOp.delete_resourcegroup(credentials, subscription_id, resource_group_name)
                 ct = datetime.datetime.now()
                 print("Date time after deployment and deletion of stack:-", ct)
+
+    if existingVPC=='true':
+        # Delete deployment with virtual network
+        DeployOp.delete_resourcegroup(credentials, subscription_id, resource_name_vnet)
+        ct = datetime.datetime.now()
+        print("Deleted the deployment which contains the virtual network:-", ct)
 
 if __name__ == '__main__':
     main(sys.argv[1], sys.argv[2], sys.argv[3], sys.argv[4], sys.argv[5], sys.argv[6], sys.argv[7], sys.argv[8], sys.argv[9], sys.argv[10], sys.argv[11], sys.argv[12])
